+++ FacturasSimplified/Invoices/InvoiceDetail.cs
@@ -11,8 +11,8 @@ public class InvoiceDetail : AuditableEntity
   public required string Description { get; set; }
 
   public int Quantity { get; set; }
-  public double UnitPrice { get; set; }
-  public double SubTotal { get; set; }
+  public decimal UnitPrice { get; set; }
+  public decimal SubTotal { get; set; }
 
   public string? SectionTitle { get; set; }
