+++ FacturasSimplified/Products/Product.cs
@@ -8,7 +8,7 @@ public class Product
   public string? Name { get; set; }
 
   public required string Description { get; set; }
-  public required double Amount { get; set; }
+  public required decimal Amount { get; set; }
 
 
   // navigation props
